ADEN-2061 moved inner class to ModalSelectorsHelper
